+++ src/NzbDrone.Core/Profiles/Metadata/MetadataProfileService.cs
@@ -280,7 +280,7 @@ namespace NzbDrone.Core.Profiles.Metadata
                 book.Editions = FilterEditions(book.Editions.Value, localEditions, localFiles, profile);
             }
 
-            FilterByPredicate(hash, x => x.ForeignBookId, localHash, profile, (x, p) => x.Editions.Value.Any(e => e.PageCount > p.MinPages) || x.Editions.Value.All(e => e.PageCount == 0), "minimum page count not met");
+            FilterByPredicate(hash, x => x.ForeignBookId, localHash, profile, (x, p) => p.MinPages <= 0 || x.Editions.Value.Any(e => e.PageCount > p.MinPages), "minimum page count not met");
             FilterByPredicate(hash, x => x.ForeignBookId, localHash, profile, (x, p) => x.Editions.Value.Any(), "all editions filtered out");
 
             return hash.ToList();